+++ input/parameters/lithium-ion/anodes/graphite_mcmb2528_Marquis2019/graphite_mcmb2528_ocp_Dualfoil1998.py
@@ -1,4 +1,4 @@
-import autograd.numpy as np
+from pybamm import exp, tanh
 
 
 def graphite_mcmb2528_ocp_Dualfoil1998(sto):
@@ -15,15 +15,15 @@ def graphite_mcmb2528_ocp_Dualfoil1998(sto):
 
     u_eq = (
         0.194
-        + 1.5 * np.exp(-120.0 * sto)
-        + 0.0351 * np.tanh((sto - 0.286) / 0.083)
-        - 0.0045 * np.tanh((sto - 0.849) / 0.119)
-        - 0.035 * np.tanh((sto - 0.9233) / 0.05)
-        - 0.0147 * np.tanh((sto - 0.5) / 0.034)
-        - 0.102 * np.tanh((sto - 0.194) / 0.142)
-        - 0.022 * np.tanh((sto - 0.9) / 0.0164)
-        - 0.011 * np.tanh((sto - 0.124) / 0.0226)
-        + 0.0155 * np.tanh((sto - 0.105) / 0.029)
+        + 1.5 * exp(-120.0 * sto)
+        + 0.0351 * tanh((sto - 0.286) / 0.083)
+        - 0.0045 * tanh((sto - 0.849) / 0.119)
+        - 0.035 * tanh((sto - 0.9233) / 0.05)
+        - 0.0147 * tanh((sto - 0.5) / 0.034)
+        - 0.102 * tanh((sto - 0.194) / 0.142)
+        - 0.022 * tanh((sto - 0.9) / 0.0164)
+        - 0.011 * tanh((sto - 0.124) / 0.0226)
+        + 0.0155 * tanh((sto - 0.105) / 0.029)
     )
 
     return u_eq