+++ pybamm/simulation.py
@@ -1,6 +1,23 @@
+#
+# Simulation class
+#
+import pickle
 import pybamm
 import numpy as np
 import copy
+
+
+def isnotebook():
+    try:
+        shell = get_ipython().__class__.__name__
+        if shell == "ZMQInteractiveShell":
+            return True  # Jupyter notebook or qtconsole
+        elif shell == "TerminalInteractiveShell":
+            return False  # Terminal running IPython
+        else:
+            return False  # Other type (?)
+    except NameError:
+        return False  # Probably standard Python interpreter
 
 
 class Simulation:
@@ -60,6 +77,12 @@ class Simulation:
 
         self.reset()
 
+        # ignore runtime warnings in notebooks
+        if isnotebook():
+            import warnings
+
+            warnings.filterwarnings("ignore")
+
     def set_defaults(self):
         """
         A method to set all the simulation specs to default values for the
@@ -77,7 +100,7 @@ class Simulation:
         """
         A method to reset a simulation back to its unprocessed state.
         """
-        self.model = self._model_class(self._model_options)
+        self.model = self.model.new_copy(self._model_options)
         self.geometry = copy.deepcopy(self._unprocessed_geometry)
         self._model_with_set_params = None
         self._built_model = None
@@ -100,13 +123,22 @@ class Simulation:
         )
         self._parameter_values.process_geometry(self._geometry)
 
-    def build(self):
+    def build(self, check_model=True):
         """
         A method to build the model into a system of matrices and vectors suitable for
         performing numerical computations. If the model has already been built or
         solved then this function will have no effect. If you want to rebuild,
         first use "reset()". This method will automatically set the parameters
         if they have not already been set.
+
+        Parameters
+        ----------
+        check_model : bool, optional
+            If True, model checks are performed after discretisation. For large
+            systems these checks can be slow, so can be skipped by setting this
+            option to False. When developing, testing or debugging it is recommened
+            to leave this option as True as it may help to identify any errors.
+            Default is True.
         """
 
         if self.built_model:
@@ -115,7 +147,9 @@ class Simulation:
         self.set_parameters()
         self._mesh = pybamm.Mesh(self._geometry, self._submesh_types, self._var_pts)
         self._disc = pybamm.Discretisation(self._mesh, self._spatial_methods)
-        self._built_model = self._disc.process_model(self._model, inplace=False)
+        self._built_model = self._disc.process_model(
+            self._model, inplace=False, check_model=check_model
+        )
 
     def solve(self, t_eval=None, solver=None):
         """
@@ -124,29 +158,41 @@ class Simulation:
 
         Parameters
         ----------
-        t_eval : numeric type (optional)
-            The times at which to compute the solution
+        t_eval : numeric type, optional
+            The times at which to compute the solution. If None the model will
+            be solved for a full discharge (1 hour / C_rate) if the discharge
+            timescale is provided. Otherwise the model will be solved up to a
+            non-dimensional time of 1.
         solver : :class:`pybamm.BaseSolver`
             The solver to use to solve the model.
         """
         self.build()
 
         if t_eval is None:
-            t_eval = np.linspace(0, 1, 100)
+            try:
+                # Try to compute discharge time
+                tau = self._parameter_values.evaluate(self.model.param.tau_discharge)
+                C_rate = self._parameter_values["C-rate"]
+                t_end = 3600 / tau / C_rate
+                t_eval = np.linspace(0, t_end, 100)
+            except AttributeError:
+                t_eval = np.linspace(0, 1, 100)
 
         if solver is None:
             solver = self.solver
 
         self._solution = solver.solve(self.built_model, t_eval)
 
-    def plot(self, quick_plot_vars=None):
+    def plot(self, quick_plot_vars=None, testing=False):
         """
         A method to quickly plot the outputs of the simulation.
 
         Parameters
         ----------
-        quick_plot_vars: list
+        quick_plot_vars: list, optional
             A list of the variables to plot.
+        testing, bool, optional
+            If False the plot will not be displayed
         """
 
         if self._solution is None:
@@ -163,7 +209,47 @@ class Simulation:
             self._solution,
             output_variables=quick_plot_vars,
         )
-        plot.dynamic_plot()
+
+        if isnotebook():
+            import ipywidgets as widgets
+
+            widgets.interact(
+                plot.plot,
+                t=widgets.FloatSlider(min=0, max=plot.max_t, step=0.05, value=0),
+            )
+        else:
+            plot.dynamic_plot(testing=testing)
+
+    def post_process_variables(self, variable_list, interp_kind="linear"):
+        """
+        Post-process the listed variables
+
+        Parameters
+        ----------
+        variable_list : list of str
+            List of the names of the variables to process
+        interp_kind : str
+            The method to use for interpolation
+
+        Returns
+        -------
+        dict
+            Dictionary of processed variables
+        """
+        if isinstance(variable_list, str):
+            variable_list = [variable_list]
+
+        variables = {}
+        for var in variable_list:
+            variables[var] = self.built_model.variables[var]
+
+        return pybamm.post_process_variables(
+            variables,
+            self._solution.t,
+            self._solution.y,
+            mesh=self._mesh,
+            interp_kind=interp_kind,
+        )
 
     @property
     def model(self):
@@ -258,23 +344,23 @@ class Simulation:
 
         Parameters
         ----------
-        model_options: dict (optional)
+        model_options: dict, optional
             A dictionary of options to tweak the model you are using
-        geometry: :class:`pybamm.Geometry` (optional)
+        geometry: :class:`pybamm.Geometry`, optional
             The geometry upon which to solve the model
-        parameter_values: dict (optional)
+        parameter_values: dict, optional
             A dictionary of parameters and their corresponding numerical
             values
-        update_parameter_values: dict (optional)
+        update_parameter_values: dict, (optional)
             A dictionary of a subset of the parameters and their corresponding
             numerical values. This will only overwrite the parameter values in
             this dictionary and leave the
-        submesh_types: dict (optional)
+        submesh_types: dict, optional
             A dictionary of the types of submesh to use on each subdomain
-        var_pts: dict (optional)
+        var_pts: dict, optional
             A dictionary of the number of points used by each spatial
             variable
-        spatial_methods: dict (optional)
+        spatial_methods: dict, optional
             A dictionary of the types of spatial method to use on each
             domain (e.g. pybamm.FiniteVolume)
         solver: :class:`pybamm.BaseSolver`
@@ -319,3 +405,23 @@ class Simulation:
             or spatial_methods
         ):
             self.reset()
+
+    def save(self, filename):
+        """Save simulation using pickle"""
+        if self.model.convert_to_format == "python":
+            # We currently cannot save models in the 'python'
+            raise NotImplementedError(
+                """
+                Cannot save simulation if model format is python.
+                Set model.convert_to_format = 'casadi' instead.
+                """
+            )
+        with open(filename, "wb") as f:
+            pickle.dump(self, f, pickle.HIGHEST_PROTOCOL)
+
+
+def load_sim(filename):
+    """Load a saved simulation"""
+    with open(filename, "rb") as f:
+        sim = pickle.load(f)
+    return sim