+++ results/2019_xx_2plus1D_pouch/shared.py
@@ -4,7 +4,9 @@ import scipy.interpolate as interp
 import matplotlib.pyplot as plt
 
 
-def make_comsol_model(comsol_variables, mesh, param, y_interp=None, z_interp=None):
+def make_comsol_model(
+    comsol_variables, mesh, param, y_interp=None, z_interp=None, thermal=True
+):
     "Make Comsol 'model' for comparison"
 
     comsol_t = comsol_variables["time"]
@@ -35,7 +37,7 @@ def make_comsol_model(comsol_variables, mesh, param, y_interp=None, z_interp=Non
                 np.column_stack((comsol_y, comsol_z)),
                 variable[:, i],
                 (grid_y, grid_z),
-                method="cubic",
+                method="linear",
             )
 
         def myinterp(t):
@@ -47,14 +49,8 @@ def make_comsol_model(comsol_variables, mesh, param, y_interp=None, z_interp=Non
     def comsol_voltage(t):
         return interp.interp1d(comsol_t, comsol_variables["voltage"])(t)
 
-    def comsol_vol_av_temperature(t):
-        return interp.interp1d(
-            comsol_t, comsol_variables["volume-averaged temperature"]
-        )(t)
-
     comsol_phi_s_cn = get_interp_fun("phi_s_cn")
     comsol_phi_s_cp = get_interp_fun("phi_s_cp")
-    comsol_temperature = get_interp_fun("temperature")
     comsol_current = get_interp_fun("current")
 
     # Create comsol model with dictionary of Matrix variables
@@ -63,32 +59,84 @@ def make_comsol_model(comsol_variables, mesh, param, y_interp=None, z_interp=Non
         "Terminal voltage [V]": comsol_voltage,
         "Negative current collector potential [V]": comsol_phi_s_cn,
         "Positive current collector potential [V]": comsol_phi_s_cp,
-        "X-averaged cell temperature [K]": comsol_temperature,
-        "Volume-averaged cell temperature [K]": comsol_vol_av_temperature,
         "Current collector current density [A.m-2]": comsol_current,
     }
+
+    # Add thermal variables
+    if thermal:
+
+        def comsol_vol_av_temperature(t):
+            return interp.interp1d(
+                comsol_t, comsol_variables["volume-averaged temperature"]
+            )(t)
+
+        comsol_temperature = get_interp_fun("temperature")
+        comsol_model.variables.update(
+            {
+                "X-averaged cell temperature [K]": comsol_temperature,
+                "Volume-averaged cell temperature [K]": comsol_vol_av_temperature,
+            }
+        )
+
     comsol_model.y_interp = y_interp
     comsol_model.z_interp = z_interp
 
     return comsol_model
 
 
-def plot_t_var(var, t, comsol_model, output_variables, param):
+def plot_t_var(
+    var, plot_times, comsol_model, output_variables, param, plot_error="both"
+):
     # discharge timescale
     tau = param.process_symbol(
         pybamm.standard_parameters_lithium_ion.tau_discharge
     ).evaluate()
 
-    plt.figure()
-    plt.plot(t, comsol_model.variables[var](t), label="COMSOL")
-    plt.plot(t, output_variables[var](t=(t / tau)), label="PyBaMM")
-    plt.xlabel(r"$t$")
-    plt.ylabel(var)
-    plt.legend()
+    pybamm_var = output_variables[var](t=(plot_times / tau))
+    comsol_var = comsol_model.variables[var](plot_times)
+
+    # Make plot
+
+    # add extra row for errors
+    if plot_error in ["abs", "rel"]:
+        n_rows = 2
+    elif plot_error == "both":
+        n_rows = 3
+    else:
+        n_rows = 1
+    fig, ax = plt.subplots(n_rows, 1, figsize=(15, 8))
+
+    ax[0].plot(plot_times, pybamm_var, "-", label="PyBaMM")
+    ax[0].plot(plot_times, comsol_var, "o", fillstyle="none", label="COMSOL")
+    if plot_error == "abs":
+        error = np.abs(pybamm_var - comsol_var)
+        ax[1].plot(plot_times, error, "-")
+    elif plot_error == "rel":
+        error = np.abs((pybamm_var - comsol_var) / comsol_var)
+        ax[1].plot(plot_times, error, "-")
+    elif plot_error == "both":
+        abs_error = np.abs(pybamm_var - comsol_var)
+        rel_error = np.abs((pybamm_var - comsol_var) / comsol_var)
+        ax[1].plot(plot_times, abs_error, "-")
+        ax[2].plot(plot_times, rel_error, "-")
+
+    # set labels
+    ax[0].set_xlabel("t")
+    ax[0].set_ylabel(var)
+    if plot_error in ["abs", "rel"]:
+        ax[1].set_xlabel("t")
+        ax[1].set_ylabel("error (" + plot_error + ")")
+    elif plot_error == "both":
+        ax[1].set_xlabel("t")
+        ax[1].set_ylabel("error (abs)")
+        ax[2].set_xlabel("t")
+        ax[2].set_ylabel("error (rel)")
+    ax[0].legend(bbox_to_anchor=(1.04, 1), loc="upper left")
+    plt.tight_layout()
 
 
 def plot_2D_var(
-    var, t, comsol_model, output_variables, param, cmap="viridis", error="abs"
+    var, t, comsol_model, output_variables, param, cmap="viridis", error="both"
 ):
     fig, ax = plt.subplots(figsize=(15, 8))
 
@@ -106,7 +154,11 @@ def plot_2D_var(
     pybamm_var = np.transpose(
         output_variables[var](y=y_plot_non_dim, z=z_plot_non_dim, t=t_non_dim)
     )
-    plt.subplot(131)
+
+    if error in ["abs", "rel"]:
+        plt.subplot(131)
+    elif error == "both":
+        plt.subplot(221)
     pybamm_plot = plt.pcolormesh(y_plot, z_plot, pybamm_var, shading="gouraud")
     plt.axis([0, y_plot[-1], 0, z_plot[-1]])
     plt.xlabel(r"$y$")
@@ -116,8 +168,14 @@ def plot_2D_var(
     plt.colorbar(pybamm_plot)
 
     # plot comsol solution
+
+    # for pos potential compute relative to voltage
     comsol_var = comsol_model.variables[var](t=t)
-    plt.subplot(132)
+
+    if error in ["abs", "rel"]:
+        plt.subplot(132)
+    elif error == "both":
+        plt.subplot(222)
     comsol_plot = plt.pcolormesh(y_plot, z_plot, comsol_var, shading="gouraud")
     plt.axis([0, y_plot[-1], 0, z_plot[-1]])
     plt.xlabel(r"$y$")
@@ -127,15 +185,47 @@ def plot_2D_var(
     plt.colorbar(comsol_plot)
 
     # plot "error"
-    plt.subplot(133)
-    if error == "abs":
-        error = np.abs(pybamm_var - comsol_var)
-    elif error == "rel":
-        error = np.abs((pybamm_var - comsol_var) / comsol_var)
-    diff_plot = plt.pcolormesh(y_plot, z_plot, error, shading="gouraud")
-    plt.axis([0, y_plot[-1], 0, z_plot[-1]])
-    plt.xlabel(r"$y$")
-    plt.ylabel(r"$z$")
-    plt.title(r"Error: " + var)
-    plt.set_cmap(cmap)
-    plt.colorbar(diff_plot)
+    if error in ["abs", "rel"]:
+        plt.subplot(133)
+        if error == "abs":
+            error = np.abs(pybamm_var - comsol_var)
+            diff_plot = plt.pcolormesh(y_plot, z_plot, error, shading="gouraud")
+        elif error == "rel":
+            error = np.abs((pybamm_var - comsol_var) / comsol_var)
+            # plot relative error up to max 10% (errors 10% and greater all take same
+            # color in plot)
+            vmax = np.min([np.max(error), 0.1])
+            diff_plot = plt.pcolormesh(
+                y_plot, z_plot, error, shading="gouraud", vmin=0, vmax=vmax
+            )
+        plt.axis([0, y_plot[-1], 0, z_plot[-1]])
+        plt.xlabel(r"$y$")
+        plt.ylabel(r"$z$")
+        plt.title(r"Error: " + var)
+        plt.set_cmap(cmap)
+        plt.colorbar(diff_plot)
+    elif error == "both":
+        plt.subplot(223)
+        abs_error = np.abs(pybamm_var - comsol_var)
+        abs_diff_plot = plt.pcolormesh(y_plot, z_plot, abs_error, shading="gouraud")
+        plt.axis([0, y_plot[-1], 0, z_plot[-1]])
+        plt.xlabel(r"$y$")
+        plt.ylabel(r"$z$")
+        plt.title(r"Error (abs): " + var)
+        plt.set_cmap(cmap)
+        plt.colorbar(abs_diff_plot)
+        plt.subplot(224)
+        scale = np.max(comsol_var) - np.min(comsol_var)
+        rel_error = np.abs((pybamm_var - comsol_var) / scale)
+        # plot relative error up to max 10% (errors 10% and greater all take same
+        # color in plot)
+        vmax = np.min([np.max(rel_error), 0.1])
+        rel_diff_plot = plt.pcolormesh(
+            y_plot, z_plot, rel_error, shading="gouraud", vmin=0, vmax=vmax
+        )
+        plt.axis([0, y_plot[-1], 0, z_plot[-1]])
+        plt.xlabel(r"$y$")
+        plt.ylabel(r"$z$")
+        plt.title(r"Error (rel): " + var)
+        plt.set_cmap(cmap)
+        plt.colorbar(rel_diff_plot)